requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
apptools==5.2.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
envisage==7.0.3
fonttools==4.47.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.60.0
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.1.1
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mayavi==4.7.4
multidict==6.0.4
numpy==1.20.0
nvidia-nccl-cu12==2.22.3
packaging==23.2
pandas==1.4.4
pillow==10.2.0
protobuf==3.20.0
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyface==8.0.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.9.1
sentry-sdk==2.7.1
setproctitle==1.3.3
Shapely==1.7.1
six==1.16.0
smmap==5.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.4
traits==6.4.3
traitsui==8.0.0
typing_extensions==4.12.2
urllib3==2.2.2
vtk==9.1.0
wandb==0.16.6
Werkzeug==3.0.1
wslink==1.12.4
xgboost==2.1.0
yarl==1.9.4
zipp==3.17.0
